refactor: csxの関数をlibsに移行、摂氏華氏変換関数を別ファイル化、utilをutilsに変更(#3)
