+++ src/components/common/CircleButton/CircleButton.tsx
@@ -1,7 +1,7 @@
 import { FC, ComponentPropsWithRef } from 'react';
 import { css } from '@emotion/react';
 import { fonts, colors } from '@/styles/constants';
-import { createDarkenColor } from '@/util/color';
+import { createDarkenColor } from '@/libs/csx';
 
 type Color = 'light' | 'dark';
 type Props = Omit<ComponentPropsWithRef<'button'>, 'color'> & {
